requirements.txt
# 核心框架
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0
litellm>=1.40.0

# 记忆
elasticsearch>=8.0.0
sentence-transformers>=3.0.0
redis>=5.0.0

# 工具
tavily-python>=0.3.0
duckduckgo-search>=6.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# API
fastapi>=0.111.0
uvicorn>=0.30.0

# 工具链
pydantic>=2.7.0
pydantic-settings>=2.3.0
loguru>=0.7.0
python-dotenv>=1.0.0
typer>=0.12.0
